Add a line mix-in class + undo changes made for histogram inheriting from curve
histogram currently broken
